Rename Standard implementation to Basic
