+++ src/OpenTracing/Basic/Propagator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenTracing\Standard;
+namespace OpenTracing\Basic;
 
 use OpenTracing;
 use OpenTracing\Extractor;
