+++ src/OpenTracing/Basic/SplitTextPropagator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenTracing\Standard;
+namespace OpenTracing\Basic;
 
 use OpenTracing;
